insurance_claims_app/requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langgraph>=0.2.0
langchain-core>=0.2.0
oci>=2.100.0
oracledb>=2.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
chromadb>=0.4.22
python-multipart>=0.0.6
httpx>=0.26.0
sentence-transformers>=2.3.0
transformers>=4.36.0
pillow>=10.0.0
torch>=2.0.0
